整体代码初步优化，支持 PG 数据库

+++ pkg/api/dto/point_log_dto.go
@@ -28,7 +28,7 @@ func toPointDto(point model.PointLog) PointDto {
 	}
 	return PointDto{
 		ID:        int64(point.ID),
-		Point:     point.Points,
+		Point:     float64(point.Points),
 		UserID:    point.UserID,
 		Nickname:  nickname,
 		Reason:    point.Reason,
